requirements.txt
djangorestframework==3.14.0
django-cors-headers==4.1.0
python-dateutil==2.8.2